Generates full path informat for R files called from python
To call the R files in the BEASTIE directory from Rscript using os.system
the full path needs to be passed to Rscript. Otherwise Rscript will look in
the current working directory -- the directory the user was in when the ran
the `beastie` command -- and will fail because the files won't be found.